refactor on boarding by removing pages const data from OnBoardingData class to the ui that uses it so doing simple changes to some other files
