+++ lib/features/onboarding/presentation/view/onboarding_screen.dart
@@ -1,19 +1,13 @@
 import 'package:aoun_tu/core/utls/colors.dart';
-import 'package:aoun_tu/features/onboarding/presentation/view/widgets/onboarding_view_body.dart';
+import 'package:aoun_tu/features/onboarding/presentation/view/widgets/on_boarding_body.dart';
 import 'package:aoun_tu/features/onboarding/presentation/view_model/onboarding_cubit.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import '../../../../injection_container.dart';
-import '../../data/on_boarding_data.dart';
 
-class OnBoardingView extends StatefulWidget {
+class OnBoardingView extends StatelessWidget {
   const OnBoardingView({Key? key}) : super(key: key);
 
-  @override
-  State<OnBoardingView> createState() => _OnBoardingViewState();
-}
-
-class _OnBoardingViewState extends State<OnBoardingView> {
   @override
   Widget build(BuildContext context) {
     return BlocProvider(
@@ -29,38 +23,5 @@ class _OnBoardingViewState extends State<OnBoardingView> {
   }
 }
 
-class OnBoardingBody extends StatefulWidget {
-  const OnBoardingBody({Key? key}) : super(key: key);
 
-  @override
-  State<OnBoardingBody> createState() => _OnBoardingBodyState();
-}
 
-class _OnBoardingBodyState extends State<OnBoardingBody> {
-  @override
-  void dispose() {
-    if (context.mounted) {
-      BlocProvider.of<OnBoardingCubit>(context).pageController.dispose();
-      super.dispose();
-    }
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    var onBoardingCubit = BlocProvider.of<OnBoardingCubit>(context);
-    return PageView.builder(
-      itemBuilder: (context, index) {
-        return OnBoardingViewBody(
-          title: pages[index].title,
-          description: pages[index].description,
-          image: pages[index].image,
-        );
-      },
-      controller: onBoardingCubit.pageController,
-      onPageChanged: (c) {
-        onBoardingCubit.getCurrentPageViewIndex(c);
-      },
-      itemCount: pages.length,
-    );
-  }
-}
